client: memory waste fix introduced by http://www.etlegacy.com/projects/etlegacy/repository/revisions/99cf888f03db3b491c1b203c59bde667e2f129fd

+++ src/client/snd_mem.c
@@ -88,7 +88,7 @@ void SND_setup(void)
 
 	cv = Cvar_Get("com_soundMegs", DEF_COMSOUNDMEGS, CVAR_LATCH | CVAR_ARCHIVE);
 
-	scs = (cv->integer * 1536);
+	scs = (cv->integer * 512); // q3 uses a value of 1536 - reverted to genuine ET value
 
 	buffer = malloc(scs * sizeof(sndBuffer));
 	// allocate the stack based hunk allocator
